+++ lib/common/emit.c
@@ -1973,7 +1973,8 @@ static void splitBSpline(bezier *bz, double t, bezier *left, bezier *right) {
  * Return non-zero if color spec is incorrect
  */
 static int multicolor(GVJ_t *job, edge_t *e, char **styles, const char *colors,
-                      double arrowsize, double penwidth) {
+                      double tailarrowsize, double headarrowsize,
+                      double penwidth) {
     bezier bz;
     bezier bz0, bz_l, bz_r;
     int rv;
@@ -2036,12 +2037,12 @@ static int multicolor(GVJ_t *job, edge_t *e, char **styles, const char *colors,
 	if (bz.sflag) {
     	    gvrender_set_pencolor(job, LIST_FRONT(&segs)->color);
     	    gvrender_set_fillcolor(job, LIST_FRONT(&segs)->color);
-	    arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0], arrowsize, penwidth, bz.sflag);
+	    arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0], tailarrowsize, penwidth, bz.sflag);
 	}
 	if (bz.eflag) {
     	    gvrender_set_pencolor(job, endcolor);
     	    gvrender_set_fillcolor(job, endcolor);
-	    arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1], arrowsize, penwidth, bz.eflag);
+	    arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1], headarrowsize, penwidth, bz.eflag);
 	}
 	if (ED_spl(e)->size > 1 && (bz.sflag || bz.eflag) && styles)
 	    gvrender_set_style(job, styles);
@@ -2356,7 +2357,7 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
     bezier bz;
     splines offspl, tmpspl;
     pointf pf0, pf1, pf2 = { 0, 0 }, pf3, *offlist, *tmplist;
-    double arrowsize, numc2, penwidth=job->obj->penwidth;
+    double arrowsize, headarrowsize, tailarrowsize, numc2, penwidth=job->obj->penwidth;
     char* p;
     bool tapered = false;
     agxbuf buf = {0};
@@ -2365,6 +2366,8 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 
     if (ED_spl(e)) {
 	arrowsize = late_double(e, E_arrowsz, 1.0, 0.0);
+	headarrowsize = late_double(e, E_arrowheadsz, arrowsize, 0.0);
+	tailarrowsize = late_double(e, E_arrowtailsz, arrowsize, 0.0);
 	color = late_string(e, E_color, "");
 
 	if (styles) {
@@ -2387,7 +2390,7 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 	}
 
 	if (numsemi && numc) {
-	    if (multicolor(job, e, styles, color, arrowsize, penwidth)) {
+	    if (multicolor(job, e, styles, color, tailarrowsize, headarrowsize, penwidth)) {
 		color = DEFAULT_COLOR;
 	    }
 	    else
@@ -2433,10 +2436,10 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 	    if (fillcolor != color)
 		gvrender_set_fillcolor(job, fillcolor);
 	    if (bz.sflag) {
-		arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0], arrowsize, penwidth, bz.sflag);
+		arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0], tailarrowsize, penwidth, bz.sflag);
 	    }
 	    if (bz.eflag) {
-		arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1], arrowsize, penwidth, bz.eflag);
+		arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1], headarrowsize, penwidth, bz.eflag);
 	    }
 	}
 	/* if more than one color - then generate parallel Béziers, one per color */
@@ -2514,7 +2517,7 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 		    }
 		}
 		arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0],
-			arrowsize, penwidth, bz.sflag);
+			tailarrowsize, penwidth, bz.sflag);
 	    }
 	    if (bz.eflag) {
 		if (color != headcolor) {
@@ -2525,7 +2528,7 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 		    }
 		}
 		arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1],
-			arrowsize, penwidth, bz.eflag);
+			headarrowsize, penwidth, bz.eflag);
 	    }
 	    free(colors);
 	    for (size_t i = 0; i < offspl.size; i++) {
@@ -2667,11 +2670,11 @@ static void emit_edge_graphics(GVJ_t * job, edge_t * e, char** styles)
 
 		if (bz.sflag) {
 		    arrow_gen(job, EMIT_TDRAW, bz.sp, bz.list[0],
-		              arrowsize, penwidth, bz.sflag);
+		              tailarrowsize, penwidth, bz.sflag);
 		}
 		if (bz.eflag) {
 		    arrow_gen(job, EMIT_HDRAW, bz.ep, bz.list[bz.size - 1],
-		              arrowsize, penwidth, bz.eflag);
+		              headarrowsize, penwidth, bz.eflag);
 		}
 		if (ED_spl(e)->size > 1 && (bz.sflag || bz.eflag) && styles)
 		    gvrender_set_style(job, styles);
@@ -4364,4 +4367,3 @@ bool findStopColor(const char *colorlist, char *clrs[2], double *frac) {
     LIST_FREE(&segs);
     return true;
 }
-